Moving away from using enums in the event
Keeping a strongly typed concept, even an enum - could cause problems
in the future if the type was to evolve into something else. This would
implicitly mean that you'd need to version even the type used on the event.
This escalates the complexity of events by a tenfold and should therefor
not be used.

Primitives only!