Looks like this fixes #532
Autofac does something that seems very strange.
It needs a bit more investigation. But it invokes something that
actually is not supposed to get invoked somehow.

An assembly provider that provides assemblies on disk - which
is not needed.

Anywho, this line of code was put in a few days ago to get self
binding working and it was not good. Need the Self registration source
thing instead.