Fix Auto-Sort and remove Refresh from Case Reports
The implementation of the refresh button didn't work (it didn't call the
backend to get new data), so it's gone. The refreshing can be done with
a normal browser refresh (which is more intuitive).

The auto-sort requested in #623 works now, but is maybe not the cleanest
way to do it. So maybe the person who built the sortable-table should
have a look and improve upon it.

Both the selected sorting and filters are lost when refreshing the page.
I think the best way to fix this is to use the Router to put those
things in the URL (really the only way to keep state).