Moved authentication UserManager to service
Moved the handling of user authentication and login-redirect from the
nav-top-bar component to an authentication service. This way it can be
re-used more easily across the app.

Also added a small component that only renders the contents inside it if
the user is logged in. Placing this outside the main content of the page
fixes the small issue where all the content flashed quickly before the
user was redirected to the login page.