Refactor Case Report table and store sorting in URL
Changed the sortable table used in case report listing. It's a bit more
generic now, and could easily be made completely generic if we want to
re-use a sortable table somewhere else in the frontends.

Also, the sorting is now stored in the URL, so that it is kept after
refreshing the page to get new data.